requirements.txt
Flask==3.0.3
flask-cors==5.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
gunicorn==23.0.0